engine: extract esm/layers/spreadsheet_layer.js
